Rtlamr update (#2)
* (maint) bump rtlamr to 0.9.4, plus other updates

This commit bumps rtlamr to 0.9.4 and removes some of the version
pinning for other packages so they're automatically updated when the
container image is built. It also bumps the hassio-addons build
container to the latest version (7.7.1), based on Debian Bookworm.

* Update Dockerfile

fix botched merge

* Normalize line endings for .sh files to LF

* Update base and Go images in Dockerfile

Bump debian-base image to 7.8.3 and golang image to 1.24.5-alpine3.22 for improved security and compatibility.

* Update base images to Debian 7.8.3 in build config

Upgrades the amd64, armv7, and aarch64 base images to version 7.8.3 and removes armhf and codenotary configuration from build.yaml.

---------

Co-authored-by: roger <133084789+parrot-tailor@users.noreply.github.com>